bugfix(resolve): in TypeScript resolve mjs, cjs to their ts variants as well
