+++ test/extract/resolve/__mocks__/resolve-to-cjs-when-imported-as-cjs/i-am-just-commonjs.cts
@@ -1,0 +1,1 @@
+// actually sorta lying
