+++ test/extract/resolve/__mocks__/resolve-to-js-even-when-imported-as-js/i-am-genuinely-javascript.ts
@@ -1,0 +1,1 @@
+// it's a LIE
