+++ test/extract/resolve/__mocks__/resolve-to-mjs-when-imported-as-mjs/i-am-just-esm.mts
@@ -1,0 +1,1 @@
+// actually sorta lying
