+++ test/extract/resolve/__mocks__/resolve-to-mts-even-when-imported-as-mjs/i-am-secretly-typescript.mts
@@ -1,0 +1,1 @@
+
